requirements.txt
# Voice Control System Requirements
# =================================
# 
# Core dependencies for voice recognition and GUI functionality
# Optimized for minimal size and better GitHub Codespaces compatibility

# Voice Recognition (Core)
openai-whisper>=20231117    # Local speech-to-text processing
numpy>=1.21.0              # Array operations for audio processing

# Audio Processing
pyaudio>=0.2.11            # Audio input/output (requires system dependencies)
scipy>=1.7.0               # Signal processing for audio filtering

# File Monitoring and GUI
watchdog>=2.1.0            # File system monitoring
matplotlib>=3.5.0          # Optional: for advanced audio visualization

# Build Tools (Optional)
pyinstaller>=5.0.0         # For building standalone executables

# Note: PyTorch is automatically installed with openai-whisper
# Note: tkinter is included with most Python installations